requirements.txt
Django==5.1.4
django-environ==0.11.2
django-ninja==1.3.0
djangorestframework==3.15.2